New common-network module (#149)
