--- common-data/src/iosMain/kotlin/dev/sergiobelda/todometer/common/data/api/TodometerApi.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright 2021 Sergio Belda
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.sergiobelda.todometer.common.data.api
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Runnable
-import platform.darwin.dispatch_async
-import platform.darwin.dispatch_get_main_queue
-import platform.darwin.dispatch_queue_t
-import kotlin.coroutines.CoroutineContext
-
-internal actual val ApplicationDispatcher: CoroutineDispatcher =
-    NsQueueDispatcher(dispatch_get_main_queue())
-
-internal class NsQueueDispatcher(
-    private val dispatchQueue: dispatch_queue_t
-) : CoroutineDispatcher() {
-    override fun dispatch(context: CoroutineContext, block: Runnable) {
-        dispatch_async(dispatchQueue) {
-            block.run()
-        }
-    }
-}
