+++ common-network/common_network.podspec
@@ -1,0 +1,39 @@
+Pod::Spec.new do |spec|
+    spec.name                     = 'common_network'
+    spec.version                  = '1.0'
+    spec.homepage                 = 'Link to the Shared Module homepage'
+    spec.source                   = { :http=> ''}
+    spec.authors                  = ''
+    spec.license                  = ''
+    spec.summary                  = 'Common network'
+    spec.vendored_frameworks      = 'build/cocoapods/framework/common-network.framework'
+    spec.libraries                = 'c++'
+    spec.ios.deployment_target = '14.1'
+                
+                
+    spec.pod_target_xcconfig = {
+        'KOTLIN_PROJECT_PATH' => ':common-network',
+        'PRODUCT_MODULE_NAME' => 'common-network',
+    }
+                
+    spec.script_phases = [
+        {
+            :name => 'Build common_network',
+            :execution_position => :before_compile,
+            :shell_path => '/bin/sh',
+            :script => <<-SCRIPT
+                if [ "YES" = "$COCOAPODS_SKIP_KOTLIN_BUILD" ]; then
+                  echo "Skipping Gradle build task invocation due to COCOAPODS_SKIP_KOTLIN_BUILD environment variable set to \"YES\""
+                  exit 0
+                fi
+                set -ev
+                REPO_ROOT="$PODS_TARGET_SRCROOT"
+                "$REPO_ROOT/../gradlew" -p "$REPO_ROOT" $KOTLIN_PROJECT_PATH:syncFramework \
+                    -Pkotlin.native.cocoapods.platform=$PLATFORM_NAME \
+                    -Pkotlin.native.cocoapods.archs="$ARCHS" \
+                    -Pkotlin.native.cocoapods.configuration="$CONFIGURATION"
+            SCRIPT
+        }
+    ]
+                
+end
