+++ common-network/src/commonMain/kotlin/dev/sergiobelda/todometer/common/network/TodometerApi.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2021 Sergio Belda
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.sergiobelda.todometer.common.network
+
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.serialization.json.Json
+
+expect val ApplicationDispatcher: CoroutineDispatcher
+
+class TodometerApi {
+    val client = HttpClient {
+        install(ContentNegotiation) {
+            json(
+                Json {
+                    prettyPrint = true
+                    isLenient = true
+                }
+            )
+        }
+    }
+
+    companion object {
+
+        private const val SCHEME = "http"
+        const val HOST = "192.168.0.11"
+        const val PORT = "8080"
+        const val ENDPOINT_URL = "$SCHEME://$HOST:$PORT"
+        const val VERSION_1 = "/v1"
+        const val TASK_LIST_PATH = "/tasklists"
+        const val TASK_PATH = "/tasks"
+        const val TAG_PATH = "/tags"
+    }
+}
