+++ settings.gradle.kts
@@ -19,6 +19,7 @@ include(":common-android-resources")
 include(":common-compose-ui")
 include(":common-data")
 include(":common-domain")
+include(":common-network")
 include(":common-preferences")
 include(":desktop")
 include(":ios")
